@@ -4,10 +4,10 @@ class VacationsBal {
   final int id;
   final String availableafter;
   final String monitortype;
-  final double usedbal;
-    final double remainderbal;
-    final double previousbal;
-    final double begbal;
+  final double? usedbal;
+    final double? remainderbal;
+    final double? previousbal;
+    final double? begbal;
     final bool allowHourTrans;
  String getAriaLabel(){
 return this.monitortype;
@@ -16,7 +16,7 @@ return this.monitortype;
  String getAriaValue(){
  return 'الرصيد المستخدم: $usedbal. , الرصيد المتبقي: $remainderbal.';
 }
-  VacationsBal({this.id,this.allowHourTrans,this.monitortype, this.availableafter, this.remainderbal, this.usedbal,this.previousbal,this.begbal});
+  VacationsBal({required this.id,required this.allowHourTrans,required this.monitortype,required this.availableafter, this.remainderbal, this.usedbal,this.previousbal,this.begbal});
 
   factory VacationsBal.fromJson(Map<String, dynamic> json) {
     return VacationsBal(
