@@ -1,8 +1,6 @@
-import 'dart:ui';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:hrapp/model/itemSelectVM.dart';
 import 'package:hrapp/model/vacations/vacationsPostVM.dart';
 import 'package:hrapp/services/smartApiService.dart';
@@ -15,15 +13,14 @@ import '../../model/vacations_bal.dart';
 import 'package:intl/intl.dart';
 
 class VacationsPostPage extends StatefulWidget {
-  const VacationsPostPage({Key key, @required this.vm, @required this.postType})
-      : super(key: key);
+  const VacationsPostPage({super.key, 
+  required this.vm, required this.postType});
 
   final VacationsBal vm;
   final String postType;
 
   @override
-  _VacationsPostPageState createState() =>
-      _VacationsPostPageState();
+  _VacationsPostPageState createState() => _VacationsPostPageState();
 }
 
 class _VacationsPostPageState extends State<VacationsPostPage> {
@@ -39,7 +36,7 @@ class _VacationsPostPageState extends State<VacationsPostPage> {
   TextEditingController _toDateInputValue = TextEditingController();
   TextEditingController _noteInput = TextEditingController();
   TextEditingController _sparEmpInput = TextEditingController();
-  ItemSelectVM _sparEmpInputValue = ItemSelectVM();
+  ItemSelectVM? _sparEmpInputValue ;
 
   //String _responseBody = '<empty>';
   //String _error = '<none>';
@@ -56,55 +53,53 @@ class _VacationsPostPageState extends State<VacationsPostPage> {
   Widget build(BuildContext context) {
     //final height = MediaQuery.of(context).size.height;
     final isTimePost = this.widget.postType == VactionPostType.Permission;
-    return 
-    
-    
-    CupertinoPageScaffold(
-
+    return CupertinoPageScaffold(
       backgroundColor: SmartAppTheme.white,
-        navigationBar: CupertinoNavigationBar(
-     backgroundColor: SmartAppTheme.white,
-    // automaticallyImplyMiddle: false,
-      border: null,
-         // iconTheme: IconThemeData(color: SmartAppTheme.iconColor),
-        ),
+      navigationBar: CupertinoNavigationBar(
+        backgroundColor: SmartAppTheme.white,
+        // automaticallyImplyMiddle: false,
+        border: null,
+        // iconTheme: IconThemeData(color: SmartAppTheme.iconColor),
+      ),
       child: Container(
-        //  height: height,
+          //  height: height,
 
-        child: SingleChildScrollView(
-          // child: Form(
-            
-          padding:EdgeInsets.symmetric(vertical: 0,horizontal: MediaQuery.of(context).size.width * .1),
-          child:Material(
-            color: Colors.white,
-            child:   Form(
-             
+          child: SingleChildScrollView(
+        // child: Form(
+
+        padding: EdgeInsets.symmetric(
+            vertical: 0, horizontal: MediaQuery.of(context).size.width * .1),
+        child: Material(
+          color: Colors.white,
+          child: Form(
               key: formKey,
               // padding: const EdgeInsets.all(16.0),
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.center,
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
-                //  const SizedBox(height: 40),
+                  //  const SizedBox(height: 40),
                   Image.asset(
                     'res/img/ic_cup.png',
                     alignment: Alignment.center,
-                    
+
                     height: 70,
-                   // scale: 2,
+                    // scale: 2,
                     width: 70,
                     excludeFromSemantics: true,
-                    color:isTimePost ? Colors.orange : SmartAppTheme.colorPrimary,
+                    color:
+                        isTimePost ? Colors.orange : SmartAppTheme.colorPrimary,
                   ),
-                ExcludeSemantics(
-                  child: Text(this.widget.vm.monitortype, style: SmartAppTheme.title),
-                ),
-                 
+                  ExcludeSemantics(
+                    child: Text(this.widget.vm.monitortype,
+                        style: SmartAppTheme.title),
+                  ),
+
                   const SizedBox(height: 5),
                   ExcludeSemantics(
-                  child: Text(isTimePost ? "اجازة ساعية" : "اجازة يومية",style: SmartAppTheme.caption),
-                ),
-                
+                    child: Text(isTimePost ? "اجازة ساعية" : "اجازة يومية",
+                        style: SmartAppTheme.caption),
+                  ),
 
                   //  const Divider(),
                   const SizedBox(height: 10),
@@ -120,31 +115,28 @@ class _VacationsPostPageState extends State<VacationsPostPage> {
                       : smartDateInput(this._toDateInput,
                           this._toDateInputValue, "الى تاريخ"),
                   const SizedBox(height: 10),
-              
-          Semantics(
-   container: true,
-   button: true,
-   onTap: ()  => selectEmp(),
-    excludeSemantics: true,
-          label:'الموظف المناوب',
-        value:  this._sparEmpInput.text,  
-        child:  TextFormField(
-          style:SmartAppTheme.defaultTextStyle ,
-                    controller: this._sparEmpInput,
-                   readOnly: true,
-                    onTap: ()  => selectEmp(),
-                    decoration: const InputDecoration(
-                      labelText: 'الموظف المناوب',
-                      
-                      border: UnderlineInputBorder(),
-                      suffixIcon: Icon(Icons.search),
-                    ),
-                    
-               
-                  )),
+
+                  Semantics(
+                      container: true,
+                      button: true,
+                      onTap: () => selectEmp(),
+                      excludeSemantics: true,
+                      label: 'الموظف المناوب',
+                      value: this._sparEmpInput.text,
+                      child: TextFormField(
+                        style: SmartAppTheme.defaultTextStyle,
+                        controller: this._sparEmpInput,
+                        readOnly: true,
+                        onTap: () => selectEmp(),
+                        decoration: const InputDecoration(
+                          labelText: 'الموظف المناوب',
+                          border: UnderlineInputBorder(),
+                          suffixIcon: Icon(Icons.search),
+                        ),
+                      )),
                   const SizedBox(height: 4),
                   TextFormField(
-                       style:SmartAppTheme.defaultTextStyle,
+                    style: SmartAppTheme.defaultTextStyle,
                     controller: this._noteInput,
                     maxLines: 2,
                     decoration: const InputDecoration(
@@ -178,7 +170,7 @@ class _VacationsPostPageState extends State<VacationsPostPage> {
         this._fromDateInputValue.text = "";
         this._toDateInputValue.text = "";
 
-        this._sparEmpInputValue = ItemSelectVM();
+        this._sparEmpInputValue = null;
       }
       // this._responseBody = '<empty>';
       // this._error = '<none>';
@@ -188,6 +180,8 @@ class _VacationsPostPageState extends State<VacationsPostPage> {
 
   Future<void> _post() async {
     final form = formKey.currentState;
+    if(form==null)
+    return;
     if (form.validate()) {
       form.save();
       _reset(resetControllers: false);
@@ -197,153 +191,93 @@ class _VacationsPostPageState extends State<VacationsPostPage> {
           ? AppUrl.PermissionTransPost
           : AppUrl.VacationTransPost;
 
-      final Map<String, dynamic> response =await postFormData(url, _getFormData());
+      final Map<String, dynamic> response =
+          await postFormData(url, _getFormData());
       if (response['success']) {
         Navigator.pop(context, true);
       } else {
-        smartErrorToast(context, "لم تتم العملية", response['message'].toString());
+        smartErrorToast(
+            context, "لم تتم العملية", response['message'].toString());
       }
       setState(() => this._pending = false);
     }
   }
 
-   selectEmp() async{
-         final result = await Navigator.push(
-                          context,
-                          CupertinoPageRoute(
-                              builder: (context) => ItemSelectPage(
-                                  title: 'الموظف المناوب',
-                                  url: AppUrl.EmployeeFilter)));
-                      if (result != null) {
-                        this._sparEmpInput.text = result.name;
-                        _sparEmpInputValue = ItemSelectVM(id: result.id, name: result.name);
-                      }
-                      // else{
-                      //    this._sparEmpInput.text="test2";
-                      // }
- }
+  selectEmp() async {
+    final result = await Navigator.push(
+        context,
+        CupertinoPageRoute(
+            builder: (context) => ItemSelectPage(
+                title: 'الموظف المناوب', url: AppUrl.EmployeeFilter)));
+    if (result != null) {
+      this._sparEmpInput.text = result.name;
+      _sparEmpInputValue = ItemSelectVM(id: result.id, name: result.name);
+    }
+    // else{
+    //    this._sparEmpInput.text="test2";
+    // }
+  }
 
   Map<String, dynamic> _getFormData() {
     return <String, dynamic>{
       'StartDate': this._fromDateInputValue.text,
       'EndDate': this._toDateInputValue.text,
       'Trans_Remark': this._noteInput.text,
-      'Employee_IdSpare': this._sparEmpInputValue.id,
+      'Employee_IdSpare': this._sparEmpInputValue?.id,
       'MonitorType_Id': this.widget.vm.id,
-
-      
     };
   }
 
   smartDateInput(TextEditingController dateInput,
           TextEditingController valueInput, String label) =>
-               Semantics(
-   container: true,
-   button: true,
-   onTap: ()  => selectDate(dateInput,valueInput),
-    excludeSemantics: true,
-          label:label,
-        value:  dateInput.text,  
-        child:
-      Center(
-          child: TextFormField(
-               style:SmartAppTheme.defaultTextStyle ,
-        validator: (value) => value.isEmpty ? "يرجى اختيار التاريخ " : null,
-        controller: dateInput,
-        decoration: InputDecoration(
-            suffixIcon: Icon(Icons.calendar_today),
-            border: UnderlineInputBorder(),
-            labelText: label),
-        readOnly: true,
-       onTap: ()  => selectDate(dateInput,valueInput),
-      )));
-
+      Semantics(
+          container: true,
+          button: true,
+          onTap: () => selectDate(dateInput, valueInput),
+          excludeSemantics: true,
+          label: label,
+          value: dateInput.text,
+          child: Center(
+              child: TextFormField(
+            style: SmartAppTheme.defaultTextStyle,
+            validator: (value) =>value==null|| value.isEmpty ? "يرجى اختيار التاريخ " : null,
+            controller: dateInput,
+            decoration: InputDecoration(
+                suffixIcon: Icon(Icons.calendar_today),
+                border: UnderlineInputBorder(),
+                labelText: label),
+            readOnly: true,
+            onTap: () => selectDate(dateInput, valueInput),
+          )));
 
   smartTimeInput(TextEditingController dateInput,
           TextEditingController valueInput, String label) =>
-                      Semantics(
-   container: true,
-   button: true,
-   onTap: ()  => selectDateTime(dateInput,valueInput),
-    excludeSemantics: true,
-          label:label,
-        value:  dateInput.text,  
-        child:
-      Center(
-          child: TextFormField(
-               style:SmartAppTheme.defaultTextStyle ,
-        validator: (value) => value.isEmpty ? "يرجى اختيار الوقت " : null,
-        controller: dateInput,
-        decoration: InputDecoration(
-            suffixIcon: Icon(Icons.access_time),
-            border: UnderlineInputBorder(),
-            labelText: label),
-        readOnly: true,
-      onTap: ()  => selectDateTime(dateInput,valueInput),
-        ),
-      ));
+      Semantics(
+          container: true,
+          button: true,
+          onTap: () => selectDateTime(dateInput, valueInput),
+          excludeSemantics: true,
+          label: label,
+          value: dateInput.text,
+          child: Center(
+            child: TextFormField(
+              style: SmartAppTheme.defaultTextStyle,
+              validator: (value) =>value==null|| value.isEmpty ? "يرجى اختيار الوقت " : null,
+              controller: dateInput,
+              decoration: InputDecoration(
+                  suffixIcon: Icon(Icons.access_time),
+                  border: UnderlineInputBorder(),
+                  labelText: label),
+              readOnly: true,
+              onTap: () => selectDateTime(dateInput, valueInput),
+            ),
+          ));
 
-
-  void selectDateTime(TextEditingController dateInput,TextEditingController valueInput) async {
-      final DateTime now = DateTime.now();
- bool hasSetDate=false;
-   await  showCupertinoModalPopup(
-        context: context,
-        builder: (_) => Container(
-              height: 380,
-              color: Color.fromARGB(255, 255, 255, 255),
-              child: Column(
-                children: [
-                  Container(
-                    height: 300,
-                    child: CupertinoDatePicker(    
-                      use24hFormat: true, 
-                             
-                mode:CupertinoDatePickerMode.time,
-                        initialDateTime: DateTime.now(),
-                        onDateTimeChanged: (pickedDate) {
-                          hasSetDate=true;
-                final DateTime formatedTimeValue = DateTime(now.year,now.month,now.day,pickedDate.hour,pickedDate.minute);
-            //String formattedDate = pickedDate.format(context);
-            String formattedDate = DateFormat('HH:mm').format(formatedTimeValue);
-            setState(() {
-              valueInput.text = DateFormat('yyyy-MM-dd HH:mm').format(formatedTimeValue);
-              dateInput.text =formattedDate; //set output date to TextField value.
-               
-            formKey.currentState.validate();
-            });
-                        }),
-                  ),
-            Divider(height: 0, thickness: 1),
-                   CupertinoButton(
-                    child: Text('موافق',style: SmartAppTheme.defaultIosTextStype,),
-                    onPressed:(){
-                      if(hasSetDate==false){
-               final DateTime formatedTimeValue = DateTime.now();
-            //String formattedDate = pickedDate.format(context);
-            String formattedDate = DateFormat('HH:mm').format(formatedTimeValue);
-            setState(() {
-              valueInput.text = DateFormat('yyyy-MM-dd HH:mm').format(formatedTimeValue);
-              dateInput.text =formattedDate; //set output date to TextField value.
-               
-            formKey.currentState.validate();
-            });
-                      };
-                       Navigator.of(context).pop();
-                    
-                    },
-                  )
-                ],
-              ),
-            ));
-  }
-
-
-// Show the modal that contains the CupertinoDatePicker
-  void selectDate(TextEditingController dateInput,TextEditingController valueInput) async{
-      int currentYear = DateTime.now().year;
-      bool hasSetDate=false;
-   await showCupertinoModalPopup(
+  void selectDateTime(
+      TextEditingController dateInput, TextEditingController valueInput) async {
+    final DateTime now = DateTime.now();
+    bool hasSetDate = false;
+    await showCupertinoModalPopup(
         context: context,
         builder: (_) => Container(
               height: 380,
@@ -353,40 +287,116 @@ class _VacationsPostPageState extends State<VacationsPostPage> {
                   Container(
                     height: 300,
                     child: CupertinoDatePicker(
-                     //  minimumDate: DateTime(currentYear -1), 
-                       // maximumDate: DateTime(currentYear + 1),
-              
-                mode:CupertinoDatePickerMode.date ,
+                        use24hFormat: true,
+                        mode: CupertinoDatePickerMode.time,
                         initialDateTime: DateTime.now(),
-                    minimumYear:currentYear -1 ,
-                    maximumYear:currentYear + 1 , //   on
                         onDateTimeChanged: (pickedDate) {
-                          hasSetDate=true;
-             String formattedDate = DateFormat('yyyy-MM-dd').format(pickedDate);
-            setState(() {
-              valueInput.text =formattedDate; //set output date to TextField value.
-              dateInput.text = formattedDate; //set output date to TextField value.
-                formKey.currentState.validate();
-             
-            });
+                          hasSetDate = true;
+                          final DateTime formatedTimeValue = DateTime(
+                              now.year,
+                              now.month,
+                              now.day,
+                              pickedDate.hour,
+                              pickedDate.minute);
+                          //String formattedDate = pickedDate.format(context);
+                          String formattedDate =
+                              DateFormat('HH:mm').format(formatedTimeValue);
+                          setState(() {
+                            valueInput.text = DateFormat('yyyy-MM-dd HH:mm')
+                                .format(formatedTimeValue);
+                            dateInput.text =
+                                formattedDate; //set output date to TextField value.
+
+                            formKey.currentState?.validate();
+                          });
                         }),
                   ),
-            Divider(height: 0, thickness: 1),
+                  Divider(height: 0, thickness: 1),
+                  CupertinoButton(
+                    child: Text(
+                      'موافق',
+                      style: SmartAppTheme.defaultIosTextStype,
+                    ),
+                    onPressed: () {
+                      if (hasSetDate == false) {
+                        final DateTime formatedTimeValue = DateTime.now();
+                        //String formattedDate = pickedDate.format(context);
+                        String formattedDate =
+                            DateFormat('HH:mm').format(formatedTimeValue);
+                        setState(() {
+                          valueInput.text = DateFormat('yyyy-MM-dd HH:mm')
+                              .format(formatedTimeValue);
+                          dateInput.text =
+                              formattedDate; //set output date to TextField value.
+
+                          formKey.currentState?.validate();
+                        });
+                      }
+                      ;
+                      Navigator.of(context).pop();
+                    },
+                  )
+                ],
+              ),
+            ));
+  }
+
+// Show the modal that contains the CupertinoDatePicker
+  void selectDate(
+      TextEditingController dateInput, TextEditingController valueInput) async {
+    int currentYear = DateTime.now().year;
+    bool hasSetDate = false;
+    await showCupertinoModalPopup(
+        context: context,
+        builder: (_) => Container(
+              height: 380,
+              color: Color.fromARGB(255, 255, 255, 255),
+              child: Column(
+                children: [
+                  Container(
+                    height: 300,
+                    child: CupertinoDatePicker(
+                        //  minimumDate: DateTime(currentYear -1),
+                        // maximumDate: DateTime(currentYear + 1),
+
+                        mode: CupertinoDatePickerMode.date,
+                        initialDateTime: DateTime.now(),
+                        minimumYear: currentYear - 1,
+                        maximumYear: currentYear + 1, //   on
+                        onDateTimeChanged: (pickedDate) {
+                          hasSetDate = true;
+                          String formattedDate =
+                              DateFormat('yyyy-MM-dd').format(pickedDate);
+                          setState(() {
+                            valueInput.text =
+                                formattedDate; //set output date to TextField value.
+                            dateInput.text =
+                                formattedDate; //set output date to TextField value.
+                            formKey.currentState?.validate();
+                          });
+                        }),
+                  ),
+                  Divider(height: 0, thickness: 1),
                   // Close the modal
                   CupertinoButton(
-                    child: Text('موافق',style: SmartAppTheme.defaultIosTextStype,),
-                    onPressed:(){
-                      if(hasSetDate==false){
-              String formattedDate = DateFormat('yyyy-MM-dd').format(DateTime.now());
-            setState(() {
-              valueInput.text =formattedDate; //set output date to TextField value.
-              dateInput.text = formattedDate; //set output date to TextField value.
-                formKey.currentState.validate();
-             
-            });
-                      };
-                       Navigator.of(context).pop();
-                    
+                    child: Text(
+                      'موافق',
+                      style: SmartAppTheme.defaultIosTextStype,
+                    ),
+                    onPressed: () {
+                      if (hasSetDate == false) {
+                        String formattedDate =
+                            DateFormat('yyyy-MM-dd').format(DateTime.now());
+                        setState(() {
+                          valueInput.text =
+                              formattedDate; //set output date to TextField value.
+                          dateInput.text =
+                              formattedDate; //set output date to TextField value.
+                          formKey.currentState?.validate();
+                        });
+                      }
+                      ;
+                      Navigator.of(context).pop();
                     },
                   )
                 ],
@@ -395,44 +405,43 @@ class _VacationsPostPageState extends State<VacationsPostPage> {
   }
 }
 
+//  selectDateTime(TextEditingController dateInput,TextEditingController valueInput)async{
+// final DateTime now = DateTime.now();
 
-  //  selectDateTime(TextEditingController dateInput,TextEditingController valueInput)async{
-  // final DateTime now = DateTime.now();
+//         TimeOfDay pickedDate = await showTimePicker(
+//           context: context,
+//           initialTime: TimeOfDay(hour: now.hour, minute: now.minute),
+//         );
 
-  //         TimeOfDay pickedDate = await showTimePicker(
-  //           context: context,
-  //           initialTime: TimeOfDay(hour: now.hour, minute: now.minute),
-  //         );
+//         if (pickedDate != null) {
+//          // 'yyyy-MM-dd HH:mm'
+//           final DateTime formatedTimeValue = DateTime(now.year,now.month,now.day,pickedDate.hour,pickedDate.minute);
+//           String formattedDate = pickedDate.format(context);
+//           setState(() {
+//             valueInput.text = DateFormat('yyyy-MM-dd HH:mm').format(formatedTimeValue);
+//             dateInput.text =formattedDate; //set output date to TextField value.
 
-  //         if (pickedDate != null) {
-  //          // 'yyyy-MM-dd HH:mm'
-  //           final DateTime formatedTimeValue = DateTime(now.year,now.month,now.day,pickedDate.hour,pickedDate.minute);
-  //           String formattedDate = pickedDate.format(context);
-  //           setState(() {
-  //             valueInput.text = DateFormat('yyyy-MM-dd HH:mm').format(formatedTimeValue);
-  //             dateInput.text =formattedDate; //set output date to TextField value.
-               
-  //           formKey.currentState.validate();
-  //           });
-  //         }
-  //  }
+//           formKey.currentState.validate();
+//           });
+//         }
+//  }
 
-  //  selectDate2(TextEditingController dateInput,TextEditingController valueInput) async{
-  // int currentYear = DateTime.now().year;
-  //         DateTime pickedDate = await showDatePicker(
-  //             context: context,
-  //             initialDate: DateTime.now(),
-  //             firstDate: DateTime(currentYear -
-  //                 1), //DateTime.now() - not to allow to choose before today.
-  //             lastDate: DateTime(currentYear + 1));
+//  selectDate2(TextEditingController dateInput,TextEditingController valueInput) async{
+// int currentYear = DateTime.now().year;
+//         DateTime pickedDate = await showDatePicker(
+//             context: context,
+//             initialDate: DateTime.now(),
+//             firstDate: DateTime(currentYear -
+//                 1), //DateTime.now() - not to allow to choose before today.
+//             lastDate: DateTime(currentYear + 1));
 
-  //         if (pickedDate != null) {
-  //           String formattedDate = DateFormat('yyyy-MM-dd').format(pickedDate);
-  //           setState(() {
-  //             valueInput.text =formattedDate; //set output date to TextField value.
-  //             dateInput.text = formattedDate; //set output date to TextField value.
-  //               formKey.currentState.validate();
-             
-  //           });
-  //         }
-  //  }
+//         if (pickedDate != null) {
+//           String formattedDate = DateFormat('yyyy-MM-dd').format(pickedDate);
+//           setState(() {
+//             valueInput.text =formattedDate; //set output date to TextField value.
+//             dateInput.text = formattedDate; //set output date to TextField value.
+//               formKey.currentState.validate();
+
+//           });
+//         }
+//  }
