@@ -1,11 +1,11 @@
-String validateEmail(String value) {
-  String _msg;
+String? validateEmail(String value) {
+  String? msg;
   RegExp regex = new RegExp(
       r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$');
   if (value.isEmpty) {
-    _msg = "Your username is required";
+    msg = "Your username is required";
   } else if (!regex.hasMatch(value)) {
-    _msg = "Please provide a valid emal address";
+    msg = "Please provide a valid emal address";
   }
-  return _msg;
+  return msg;
 }
