@@ -12,8 +12,9 @@ return ' الفترة: $period. , نموذج التقييم : $evaldoc. , الق
  return 'النسبة: $pg %. , الوزن: $weight %. الدرجة المحققة: $pgw %.';
 }
 
-  RecivedEvalPanal({int id, String emp, String department, String job, String evaldoc,String period,double pg,double pgw,double weight,String strongth,String weakness, String advice})
+  RecivedEvalPanal({required int id,required String emp,required String department, required String job,required String evaldoc,required String period,required double pg,required double pgw,required double weight,required String strongth,required String weakness, required String advice})
 : super(id, emp, department, evaldoc,job,period,pg,pgw,weight, strongth, weakness,  advice);
+ 
   factory RecivedEvalPanal.fromJson(Map<String, dynamic> json) {
     return RecivedEvalPanal(
       id: json['id'],
@@ -26,9 +27,9 @@ return ' الفترة: $period. , نموذج التقييم : $evaldoc. , الق
       period: json['period']??'غير محدد',
 
       
-      pg: json['pg'],
-      pgw: json['pgw'],
-      weight: json['w'],
+      pg: json['pg']??0,
+      pgw: json['pgw']??0,
+      weight: json['w']??0,
       strongth: json['strongth']??'غير محدد',
       weakness: json['weakness']??'غير محدد',
       advice: json['advice']??'غير محدد',
