@@ -23,7 +23,7 @@ return ' الموظف: $emp. , نوع الاذن $monitortype. ,  بتاريخ :
  String getAriaValue(){
  return 'ملاحظات: $note. , حالة الطلب : $appreoveState. , $rejectNote';
 }
-  ManagerPermissionPanel({this.id,this.monitortype, this.emp,this.spareEmp, this.bal, this.period,this.atDate,this.fromHour,this.toHour,this.hr_appreoved, this.manager_appreoved, this.appreoved, this.appreoveState,this.rejectNote,this.note});
+  ManagerPermissionPanel({required this.id,required this.monitortype, required this.emp,required this.spareEmp,required this.bal, required this.period,required this.atDate,required this.fromHour,required this.toHour,required this.hr_appreoved, required this.manager_appreoved,required this.appreoved,required this.appreoveState,required this.rejectNote,required this.note});
 
   factory ManagerPermissionPanel.fromJson(Map<String, dynamic> json) {
     return ManagerPermissionPanel(
@@ -32,19 +32,19 @@ return ' الموظف: $emp. , نوع الاذن $monitortype. ,  بتاريخ :
       spareEmp: json['spare_emp']??'غير محدد ',
         emp: json['n']??'غير محدد ',
       monitortype: json['monitortype']??'غير محدد ',
-      period: json['vdays'],
-      bal: json['bal'],
+      period: json['vdays']??0,
+      bal: json['bal']??0,
       fromHour: json['fhour']??'غير محدد ',
       toHour: json['tohour']??'غير محدد ',
       atDate: json['fdate']??'غير محدد ',
 
-      hr_appreoved:json['appreoved'],
-      manager_appreoved:json['manager_appreoved'],
-      appreoved:json['appreoved']&&json['manager_appreoved'],
+      hr_appreoved:json['appreoved']??false,
+      manager_appreoved:json['manager_appreoved']??false,
+      appreoved:json['appreoved']==true&&json['manager_appreoved']==true,
       
 
-appreoveState:json['appreoved']&&json['manager_appreoved']?"تمت الموافقة على الطلب":
-json['manager_appreoved']? "تم الموافقة من المدير المباشر ، في انتظار الاعتماد "
+appreoveState:json['appreoved']==true&&json['manager_appreoved']==true?"تمت الموافقة على الطلب":
+json['manager_appreoved']==true? "تم الموافقة من المدير المباشر ، في انتظار الاعتماد "
 : " قيد الانتظار للموافقة",
 
 rejectNote: json['rej_reason']??'',
