@@ -24,7 +24,7 @@ class ManagerVacTransPostVM {
 //  String getAriaLabel(){
 // return this.monitortype;
 // }
-  ManagerVacTransPostVM({this.id,this.monitortype,this.emp, this.spareEmp, this.bal, this.period,this.fromDate,this.toDate, this.appreoved, this.note});
+  ManagerVacTransPostVM({required this.id,required this.monitortype,required this.emp,required this.spareEmp,required this.bal,required this.period,required this.fromDate,required this.toDate, required this.appreoved,required this.note});
 
 //   factory ManagerVacTransPostVM.fromJson(Map<String, dynamic> json) {
 //     return ManagerVacTransPostVM(
