@@ -1,80 +1,32 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:hrapp/model/eval/department_period_eval_panal.dart';
-import 'package:hrapp/ui/eval/recivedHistoryEvalPage.dart';
 import 'package:hrapp/ui/widget/commonWidget.dart';
 import 'package:hrapp/util/app_url.dart';
 import '../../services/smartApiService.dart';
 import 'package:hrapp/ui/widget/AppTheme.dart';
+import '../widget/list_view.dart';
 
 
-class DepartmentPeriodEvalPage extends StatefulWidget {
- const DepartmentPeriodEvalPage({Key key}) : super(key: key);
 
-  @override
-  State<StatefulWidget> createState() {
-    return _DepartmentPeriodEvalPageState();
-  }
-}
-class _DepartmentPeriodEvalPageState extends State<DepartmentPeriodEvalPage> {
-ApiListResults<DepartmentPeriodEvalPanal> response;
+class DepartmentPeriodEvalPage extends StatelessWidget {
+ const DepartmentPeriodEvalPage({super.key});
 
-Future _getData() {
-   return fetchPanelData(AppUrl.DepartmentPeriodEvalPanal,(row)=>new DepartmentPeriodEvalPanal.fromJson(row))
-  .then((_response) {
-      if (mounted) {
-    setState(() {
-      response = _response;
-    });
-}
-  });
-}
 
-@override
-void initState() {
-super.initState();
-  this._getData();
-}
+
   @override
   Widget build(BuildContext context) {
-return RefreshIndicator(
-    onRefresh: _getData,
-    child: getCurrentView(context));
-  
+   return NgListView(
+      dataLoader: (context) =>
+      fetchPanelData(AppUrl.DepartmentPeriodEvalPanal,
+      (row)=>new DepartmentPeriodEvalPanal.fromJson(row)) ,
+      itemBuilder: (context,row, index,loadDataFun) =>
+      _buildRow(context, row,loadDataFun) ,
+    );
+
   }
 
-  Widget getCurrentView(BuildContext context) {
-
-
-     if (response!=null) {
-          if(!response.success){
-          return  errorView(response.message);
-
-          }
-          List<DepartmentPeriodEvalPanal> data = response.data;
-          if(data.length==0){
-          return  noResultViewView();
-          }else{
-          return _smartListView(context,data);
-          }
-          }
-        //  else if (snapshot.hasError) {
-        //   return  errorView(snapshot.error);
-        // }
-         return loadingView();
-      
- 
-  }
-
-  ListView _smartListView(BuildContext context,List<DepartmentPeriodEvalPanal> data) {
- 
-    return ListView.builder(
-      
-        itemCount: data.length,
-         scrollDirection: Axis.vertical,
-           padding: EdgeInsets.only( left:4,right: 4,bottom: 62 + MediaQuery.of(context).padding.bottom, ),
-        itemBuilder: (context, index) {
-          return _ListRowView(data: data[index],
+  Widget _buildRow(BuildContext context,DepartmentPeriodEvalPanal row,VoidCallback loadDataFun) {
+   return _ListRowView(data: row,
            callback: () {
          //    final row=data[index];
       //  Navigator.push(context, CupertinoPageRoute(builder: (context) => RecivedHistoryEvalPage(period_id:row.id,period_n: row.period )));
@@ -82,8 +34,8 @@ return RefreshIndicator(
           // var dd=data[index];
          //  print(dd.monitortype);
           });
-        });
   }
+
 
 
  
@@ -95,8 +47,8 @@ class _ListRowView extends StatelessWidget {
 
   final DepartmentPeriodEvalPanal data;
 
-  const _ListRowView({Key key, this.data,this.callback})
-      : super(key: key);
+   const _ListRowView({required this.data, required this.callback});
+
   final VoidCallback callback;
 
   @override
