@@ -1,37 +1,42 @@
-import 'package:flushbar/flushbar.dart';
+// import 'package:flushbar/flushbar.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'dart:math' as math;
 import 'AppTheme.dart';
+import 'snackbar.dart';
 
 //errorView(String msg) => Text("${msg}");
-loadingView()=> Center(child:CircularProgressIndicator(
- semanticsLabel: "جاري تحميل البيانات .. يرجى الانتظار",
-
-),);
+loadingView() => Center(
+      child: CircularProgressIndicator(
+        semanticsLabel: "جاري تحميل البيانات .. يرجى الانتظار",
+      ),
+    );
 //noResultViewView() => Center(child: Text("no result"));
 
+Widget noResultViewView() {
+  return ListView(children: [
+    Center(
+        child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            crossAxisAlignment: CrossAxisAlignment.center,
+            children: [
+          Image.asset(
+            'res/img/bg_no_item_cactus.png',
+            alignment: Alignment.center,
+            semanticLabel: "لا يوجد اي نتيجة",
+          ),
+        ]))
+  ]);
+}
 
-  Widget noResultViewView() {
-     
-     return ListView( children:[ Center(child:Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-         crossAxisAlignment: CrossAxisAlignment.center,
-       children:[
-     Image.asset('res/img/bg_no_item_cactus.png',alignment: Alignment.center,
-     semanticLabel: "لا يوجد اي نتيجة",
-     ),
-          ] ))]
-    );
-  }
-  Widget errorView(String msg) {
-     
-     return  ListView( children:[ Column(
-       children:[
+Widget errorView(String msg) {
+  return ListView(children: [
+    Column(children: [
       Container(
-    padding: EdgeInsets.symmetric(vertical: 15,horizontal: 15),
-      alignment: Alignment.center,
-      decoration: BoxDecoration(
-         // borderRadius: BorderRadius.all(Radius.circular(10)),
+        padding: EdgeInsets.symmetric(vertical: 15, horizontal: 15),
+        alignment: Alignment.center,
+        decoration: BoxDecoration(
+          // borderRadius: BorderRadius.all(Radius.circular(10)),
           boxShadow: [
             BoxShadow(
                 color: Colors.grey,
@@ -39,192 +44,224 @@ loadingView()=> Center(child:CircularProgressIndicator(
                 blurRadius: 5,
                 spreadRadius: 2)
           ],
-         color: Colors.red,
-              ),
-      child: Text( msg,
-      semanticsLabel: "حدث خطأ غير متوقع",
-      style: TextStyle(fontSize: 16, color: Colors.white,fontFamily: SmartAppTheme.fontName),
+          color: Colors.red,
+        ),
+        child: Text(
+          msg,
+          semanticsLabel: "حدث خطأ غير متوقع",
+          style: TextStyle(
+              fontSize: 16,
+              color: Colors.white,
+              fontFamily: SmartAppTheme.fontName),
+        ),
       ),
-          ),] 
-    )]);
-  }
-
-
-cardBoxDecoration()=>
- BoxDecoration(
-                color: SmartAppTheme.cardBackground,
-              //   backgroundBlendMode: BlendMode.dstOver,
-                  borderRadius: BorderRadius.only(
-                      topRight: Radius.circular(8.0),
-                      bottomRight: Radius.circular(8.0),
-                      bottomLeft: Radius.circular(8.0),
-                      topLeft: Radius.circular(68.0)),
-                  boxShadow:[
-                    BoxShadow(
-                        color: SmartAppTheme.grey.withOpacity(0.2),
-                        offset: Offset(1.1, 1.1),
-                        blurRadius: 5.0),
-                  ],
-                );
-
-cardFlatBoxDecoration()=>
- BoxDecoration(
-                color: SmartAppTheme.cardBackground,
-              //   backgroundBlendMode: BlendMode.dstOver,
-                  borderRadius: BorderRadius.only(
-                      topRight: Radius.circular(8.0),
-                      bottomRight: Radius.circular(8.0),
-                      bottomLeft: Radius.circular(8.0),
-                      topLeft: Radius.circular(8.0)),
-                  boxShadow:[
-                    BoxShadow(
-                        color: SmartAppTheme.grey.withOpacity(0.2),
-                        offset: Offset(1.1, 1.1),
-                        blurRadius: 2.0),
-                  ],
-                );
-
-
-
-smartErrorToast(BuildContext context,String title,String msg)=> Flushbar( title: title,message:msg,duration: Duration(seconds: 3),backgroundColor: Colors.redAccent,).show(context);
-smartSuccessToast(BuildContext context,String title,String msg)=> Flushbar( title: title,message:msg,duration: Duration(seconds: 3),backgroundColor: Colors.blue,).show(context);
-
-
-    smartConfirmDialog(BuildContext context,String title,String msg,Function okFn) {
-    showDialog(
-      context: context,
-      
-      builder: (context) => AlertDialog(
-      title:Text(title),
-       content:Text(msg,textAlign: TextAlign.center),
-       actions: [
-         TextButton(onPressed: okFn, child: const Text('موافق')),
-
-       ],
-
-      ),
-    );
+    ])
+  ]);
 }
+
+cardBoxDecoration() => BoxDecoration(
+      color: SmartAppTheme.cardBackground,
+      //   backgroundBlendMode: BlendMode.dstOver,
+      borderRadius: BorderRadius.only(
+          topRight: Radius.circular(8.0),
+          bottomRight: Radius.circular(8.0),
+          bottomLeft: Radius.circular(8.0),
+          topLeft: Radius.circular(68.0)),
+      boxShadow: [
+        BoxShadow(
+            color: SmartAppTheme.grey.withOpacity(0.2),
+            offset: Offset(1.1, 1.1),
+            blurRadius: 5.0),
+      ],
+    );
+
+cardFlatBoxDecoration() => BoxDecoration(
+      color: SmartAppTheme.cardBackground,
+      //   backgroundBlendMode: BlendMode.dstOver,
+      borderRadius: BorderRadius.only(
+          topRight: Radius.circular(8.0),
+          bottomRight: Radius.circular(8.0),
+          bottomLeft: Radius.circular(8.0),
+          topLeft: Radius.circular(8.0)),
+      boxShadow: [
+        BoxShadow(
+            color: SmartAppTheme.grey.withOpacity(0.2),
+            offset: Offset(1.1, 1.1),
+            blurRadius: 2.0),
+      ],
+    );
+
+smartErrorToast(BuildContext context, String title, String msg) {
+  // => Flushbar(
+  //     title: title,
+  //     message: msg,
+  //     duration: Duration(seconds: 3),
+  //     backgroundColor: Colors.redAccent,
+  //   ).show(context);
+// https://stackoverflow.com/questions/55382651/use-snackbar-with-cupertinopagescaffold
+  // ScaffoldMessenger.of(context)
+  //          .showSnackBar(SnackBar(content: Text('$title:$msg')));
+           showCupertinoSnackBar(context:context,
+           backgroundColor: CupertinoColors.systemRed,
+
+           message: '$title:$msg');
+
+// showCupertinoModalPopup(context: context, builder: (context) => SnackBar(content: Text('$title:$msg')),);
+}
+smartSuccessToast(BuildContext context, String title, String msg) {
+  //  Flushbar(
+  //     title: title,
+  //     message: msg,
+  //     duration: Duration(seconds: 3),
+  //     backgroundColor: Colors.blue,
+  //   ).show(context);
+  //  ScaffoldMessenger.of(context)
+  //         .showSnackBar(SnackBar(content: Text('$title:$msg')));
+showCupertinoSnackBar(context:context,
+backgroundColor: CupertinoColors.systemBlue,
+message: '$title:$msg');
+}
+
+smartConfirmDialog(
+    BuildContext context, String title, String msg,void Function() okFn) {
+  showDialog(
+    context: context,
+    builder: (context) => AlertDialog(
+      title: Text(title),
+      content: Text(msg, textAlign: TextAlign.center),
+      actions: [
+        TextButton(onPressed: okFn, child: const Text('موافق')),
+      ],
+    ),
+  );
+}
+
 class DialogOptions {
   String key;
   String label;
   bool checked;
-   bool reject;
-    DialogOptions({this.key, this.label,this.checked,this.reject});
-
-  }
-
-  void  smartOptionsDialog(BuildContext context,String title,List<DialogOptions> options,Function(String key) callbackFn) {
-    showDialog(
-      context: context,      
-      builder: (context) => 
-      SimpleDialog(
-      title:Text(title),
-     //  content:Text(msg,textAlign: TextAlign.center),
-     children:options.map((option) =>  ListTile(title:Text(option.label),onTap:()=>Navigator.pop(context,option.key))).toList(),
-
-      ),
-    ).then((value) => {
-      if(value!=null){
-        callbackFn(value)
-      }
-    });
+  bool reject;
+  DialogOptions({required this.key,required this.label, required this.checked,required this.reject});
 }
 
-smartStateDialog(BuildContext context,String title,List<DialogOptions> options,final List<Widget> children) {
-    showModalBottomSheet(
-      context: context,  
-          
-      builder: (context) => 
-      ListView(
-        
-    //  title:Text(title),
-     //  content:Text(msg,textAlign: TextAlign.center),
-     children:[...options.map((option) =>  ListTile(title:Text(option.label),leading:option.reject?const Icon(Icons.close,color:Colors.red): option.checked?const Icon(Icons.check_circle,color:Colors.blue):const Icon(Icons.access_time,color: Colors.orange),)).toList(),...children],
-
-      ),
-    );
+void smartOptionsDialog(BuildContext context, String title,
+    List<DialogOptions> options, Function(String key) callbackFn) {
+  showDialog(
+    context: context,
+    builder: (context) => SimpleDialog(
+      title: Text(title),
+      //  content:Text(msg,textAlign: TextAlign.center),
+      children: options
+          .map((option) => ListTile(
+              title: Text(option.label),
+              onTap: () => Navigator.pop(context, option.key)))
+          .toList(),
+    ),
+  ).then((value) => {
+        if (value != null) {callbackFn(value)}
+      });
 }
 
-
-  Widget navBackButton(BuildContext context) {
-    return InkWell(
-      onTap: () {
-        Navigator.pop(context);
-      },
-      child: Container(
-        padding: EdgeInsets.symmetric(horizontal: 10),
-        child: Row(
-          children: <Widget>[
-            Container(
-              padding: EdgeInsets.only(left: 0, top: 10, bottom: 10),
-              child: Icon(Icons.keyboard_arrow_left, color: Colors.black),
-            ),
-            Text('Back', style: TextStyle(fontSize: 12, fontWeight: FontWeight.w500))
-          ],
-        ),
-      ),
-    );
-    
-
-  }
-
-smartOptionsBottomSheet(BuildContext context,String title,List<DialogOptions> options,Function(String key) callbackFn) {
-    showModalBottomSheet(
-      context: context,      
-      builder: (context) => 
-      ListView(
-     // title:Text(title),
-     //  content:Text(msg,textAlign: TextAlign.center),
-     children:options.map((option) =>  ListTile(title:Text(option.label),onTap:()=>Navigator.pop(context,option.key))).toList(),
-
-      ),
-    ).then((value) => {
-      if(value!=null){
-        callbackFn(value)
-      }
-    });
+smartStateDialog(BuildContext context, String title,
+    List<DialogOptions> options, final List<Widget> children) {
+  showModalBottomSheet(
+    context: context,
+    builder: (context) => ListView(
+      //  title:Text(title),
+      //  content:Text(msg,textAlign: TextAlign.center),
+      children: [
+        ...options
+            .map((option) => ListTile(
+                  title: Text(option.label),
+                  leading: option.reject
+                      ? const Icon(Icons.close, color: Colors.red)
+                      : option.checked
+                          ? const Icon(Icons.check_circle, color: Colors.blue)
+                          : const Icon(Icons.access_time, color: Colors.orange),
+                ))
+            .toList(),
+        ...children
+      ],
+    ),
+  );
 }
+
+Widget navBackButton(BuildContext context) {
+  return InkWell(
+    onTap: () {
+      Navigator.pop(context);
+    },
+    child: Container(
+      padding: EdgeInsets.symmetric(horizontal: 10),
+      child: Row(
+        children: <Widget>[
+          Container(
+            padding: EdgeInsets.only(left: 0, top: 10, bottom: 10),
+            child: Icon(Icons.keyboard_arrow_left, color: Colors.black),
+          ),
+          Text('Back',
+              style: TextStyle(fontSize: 12, fontWeight: FontWeight.w500))
+        ],
+      ),
+    ),
+  );
+}
+
+smartOptionsBottomSheet(BuildContext context, String title,
+    List<DialogOptions> options, Function(String key) callbackFn) {
+  showModalBottomSheet(
+    context: context,
+    builder: (context) => ListView(
+      // title:Text(title),
+      //  content:Text(msg,textAlign: TextAlign.center),
+      children: options
+          .map((option) => ListTile(
+              title: Text(option.label),
+              onTap: () => Navigator.pop(context, option.key)))
+          .toList(),
+    ),
+  ).then((value) => {
+        if (value != null) {callbackFn(value)}
+      });
+}
+
 class CurvePainter extends CustomPainter {
   final double angle;
   //final List<Color> colors;
 
-  CurvePainter({ this.angle = 140});
+  CurvePainter({this.angle = 140});
 
   @override
   void paint(Canvas canvas, Size size) {
-    
-   List<Color> colors =
-   this.angle<=180?[
-  Colors.red,
-  Colors.red,
-   ]:
-    this.angle<=260?[
-  Colors.yellow,
-    Colors.yellow,
-
-   ]:
-    this.angle<=320?[
-  Colors.blue,
-    Colors.blue,
-
-   ]:
-    [
-                                           
-                                            Colors.green,
-                                        //   Colors.blue,
-                                             Colors.green,
-                                            // Color(0xFF8A98E8),
-                                            // Color(0xFF8A98E8),
-                                            // SmartAppTheme.nearlyDarkBlue
-
-                                          ];
+    List<Color> colors = this.angle <= 180
+        ? [
+            Colors.red,
+            Colors.red,
+          ]
+        : this.angle <= 260
+            ? [
+                Colors.yellow,
+                Colors.yellow,
+              ]
+            : this.angle <= 320
+                ? [
+                    Colors.blue,
+                    Colors.blue,
+                  ]
+                : [
+                    Colors.green,
+                    //   Colors.blue,
+                    Colors.green,
+                    // Color(0xFF8A98E8),
+                    // Color(0xFF8A98E8),
+                    // SmartAppTheme.nearlyDarkBlue
+                  ];
     // if (colors = null) {
     //   colorsList = colors ?? [];
     // } else {
     //   colorsList.addAll([Colors.white, Colors.white]);
     // }
-     // colorsList.addAll([Colors.white, Colors.white]);
+    // colorsList.addAll([Colors.white, Colors.white]);
 //colorsList=colors;
     // final shdowPaint = new Paint()
     //   ..color = Colors.black.withOpacity(0.4)
@@ -308,11 +345,11 @@ class CurvePainter extends CustomPainter {
     //  canvas.translate(centerToCircle, centerToCircle);
     //  canvas.rotate(degreeToRadians(angle+ 2));
 
-  //   canvas.save();
+    //   canvas.save();
     //  canvas.translate(0.0, -centerToCircle + 14 / 2);
-  // canvas.drawCircle(new Offset(0, 0), 14 / 5, cPaint);
-  //   canvas.restore();
-   // canvas.restore();
+    // canvas.drawCircle(new Offset(0, 0), 14 / 5, cPaint);
+    //   canvas.restore();
+    // canvas.restore();
     // canvas.restore();
   }
 
@@ -327,5 +364,10 @@ class CurvePainter extends CustomPainter {
   }
 }
 //,MainAxisAlignment mainAxisAlignment,CrossAxisAlignment crossAxisAlignment
-    //      MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
-    // CrossAxisAlignment crossAxisAlignment = CrossAxisAlignment.center,
+//      MainAxisAlignment mainAxisAlignment = MainAxisAlignment.start,
+// CrossAxisAlignment crossAxisAlignment = CrossAxisAlignment.center,
+
+
+
+
+
