@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-MaterialButton longButtons(String title, Function fun,
-    {Color color: const Color(0xfff063057), Color textColor: Colors.white}) {
+MaterialButton longButtons(String title, VoidCallback fun,
+    {Color color= const Color(0xfff063057), Color textColor= Colors.white}) {
   return MaterialButton(
     onPressed: fun,
     textColor: textColor,
